user-board 관계설정

+++ app/src/boards/boards.service.ts
@@ -4,6 +4,7 @@ import { CreateBoardDto } from './dto/create-board.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Board } from './board.entity';
 import { Repository } from 'typeorm';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class BoardsService {
@@ -12,16 +13,21 @@ export class BoardsService {
     private boardRepository: Repository<Board>,
   ) {}
 
-  async getAllBoards(): Promise<Board[]> {
-    const boards: Board[] = await this.boardRepository.find();
+  async getAllBoards(user: User): Promise<Board[]> {
+    // const boards: Board[] = await this.boardRepository.find();
+    // return boards;
+    const query = this.boardRepository.createQueryBuilder('board');
+    query.where('board.userId = :userId', { userId: user.id });
 
+    const boards = await query.getMany();
     return boards;
   }
 
-  async getBoardById(id: number): Promise<Board> {
+  async getBoardById(id: number, user: User): Promise<Board> {
     const found = await this.boardRepository.findOne({
       where: {
         id: id,
+        userId: user.id,
       },
     });
 
@@ -32,7 +38,10 @@ export class BoardsService {
     return found;
   }
 
-  async createBoard(createBoardDto: CreateBoardDto): Promise<Board> {
+  async createBoard(
+    createBoardDto: CreateBoardDto,
+    user: User,
+  ): Promise<Board> {
     const { title, description } = createBoardDto;
 
     // same as const createBoardDto = new CreateBoardDto(); createBoardDto.title = "Timber"; createBoardDto.description = "Saw";
@@ -40,23 +49,30 @@ export class BoardsService {
       title,
       description,
       status: BoardStatus.PUBLIC,
+      user,
     });
 
     await this.boardRepository.save(board);
 
+    delete board.user;
+
     return board;
   }
 
-  async deleteBoard(id: number): Promise<void> {
-    const result = await this.boardRepository.delete(id);
+  async deleteBoard(id: number, user: User): Promise<void> {
+    const result = await this.boardRepository.delete({ id, userId: user.id });
 
     if (result.affected === 0) {
       throw new NotFoundException(`Can't find Baord with id ${id}`);
     }
   }
 
-  async updateBoardStatus(id: number, status: BoardStatus): Promise<Board> {
-    const board = await this.getBoardById(id);
+  async updateBoardStatus(
+    id: number,
+    status: BoardStatus,
+    user: User,
+  ): Promise<Board> {
+    const board = await this.getBoardById(id, user);
     board.status = status;
 
     await this.boardRepository.save(board);
